% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_colors.R
\docType{data}
\name{mexico_palettes}
\alias{mexico_palettes}
\title{Mexican politics-inspired color palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
mexico_palettes
}
\description{
Use \code{\link{mexico_palette}} to produce desired color palette
}
\keyword{datasets}
